+++ lib/db/models/LockModel.ts
@@ -51,10 +51,12 @@ export class LockModel {
 
         debug("Registered model with schema.");
     }
-
+    
     public async getLock(name: string, timeout: number = 25000): Promise<boolean> {
 
-        const lock = await this.model.findOneAndUpdate({
+        let lock;
+        try {
+          lock = await this.model.findOneAndUpdate({
             name,
             timestamp : { $lte : moment().valueOf() },
           }, {
@@ -63,50 +65,80 @@ export class LockModel {
           }, {
             upsert: true,
             new: true,
-        });
-
-        if (lock && lock.instanceId === this.instanceId) {
-            this.metrics.inc("mongo_lock_hit");
-            return true;
-        } else {
+          });
+        } catch (error) {
+    
+          // duplicated key, happens when multiple clients access resource at the same time
+          if (error.code === 11000) {
             this.metrics.inc("mongo_lock_miss");
             return false;
+          }
+    
+          // other error
+          throw error;
         }
-    }
-
-    public async extendLock(name: string, extendFor: number = 25000): Promise<boolean> {
-
-        const lock = await this.model.findOneAndUpdate({
+    
+        if (lock && lock.instanceId === this.instanceId) {
+            this.metrics.inc("mongo_lock_hit");
+          return true;
+        } else {
+            this.metrics.inc("mongo_lock_miss");
+          return false;
+        }
+      }
+    
+      public async extendLock(name: string, extendFor: number = 25000): Promise<boolean> {
+    
+        let lock;
+        try {
+          lock = await this.model.findOneAndUpdate({
             name,
             timestamp : { $gte : moment().valueOf() },
             instanceId: this.instanceId,
           }, {
-            timestamp: {
-                $inc: extendFor,
+            $inc: {
+              timestamp: extendFor,
             },
           }, {
             new: true,
-        });
-
-        if (lock) {
-            this.metrics.inc("mongo_lock_extend_hit");
-            return true;
-        } else {
+          });
+        } catch (error) {
+    
+          // duplicated key, happens when multiple clients access resource at the same time
+          if (error.code === 11000) {
             this.metrics.inc("mongo_lock_extend_miss");
             return false;
+          }
+    
+          // other error
+          throw error;
         }
-    }
-
-    public async removeLock(name: string): Promise<void> {
-
-        await this.model.deleteOne({
-            name,
-            instanceId: this.instanceId,
-            timestamp : { $gte : moment().valueOf() },
+    
+        if (lock) {
+            this.metrics.inc("mongo_lock_extend_hit");
+          return true;
+        } else {
+            this.metrics.inc("mongo_lock_extend_miss");
+          return false;
+        }
+      }
+    
+      public async removeLock(name: string): Promise<any> {
+    
+        const { n } = await this.model.deleteOne({
+          name,
+          instanceId: this.instanceId,
+          timestamp : { $gte : moment().valueOf() },
         });
 
-        this.metrics.inc("mongo_lock_removed");
-    }
+        if(n === 1){
+            this.metrics.inc("mongo_lock_removed");
+            return true;
+        } else {
+            this.metrics.inc("mongo_lock_remove_failed");
+            return false;
+        }
+      }
 
     public delete(name: string) {
         return this.model.deleteMany({name}).exec();